+++ app/routes.php
@@ -1,6 +1,5 @@
 <?php
 
-// Auth routes
 use App\Controllers\AdminController;
 use App\Controllers\Auth\LoginController;
 use App\Controllers\Auth\PasswordRecoveryController;
@@ -49,7 +48,7 @@ $app->group('', function (RouteCollectorProxy $group) {
         $group->get('/{id}/edit', [UserController::class, 'edit'])->setName('user.edit');
         $group->post('/{id}', [UserController::class, 'update'])->setName('user.update');
         $group->get('/{id}/delete', [UserController::class, 'delete'])->setName('user.delete');
-        $group->get('/{id}/clear', [MediaController::class, 'clearUserMedia'])->setName('user.clear');
+        $group->get('/{id}/clear', [UserController::class, 'clearUserMedia'])->setName('user.clear');
     })->add(AdminMiddleware::class);
 
     $group->get('/profile', [ProfileController::class, 'profile'])->setName('profile');
